requirements.txt
pytest
pyelftools
psutil
loguru